Move to net.neoforged package and update license headers
